Chore : 아이콘 이미지 추가
