+++ .eslintrc.js
@@ -20,7 +20,7 @@ module.exports = {
   },
   plugins: ['react', 'prettier'],
   rules: {
-    'react/jsx-boolean-value': { always: ['personal'] },
+    'react/jsx-boolean-value': 0,
     'react/prop-types': 'off',
     'react/jsx-uses-react': 'off',
     'react/react-in-jsx-scope': 'off',
