+++ src/styles/theme.js
@@ -9,10 +9,12 @@ const fontSizes = {
 const colors = {
   colorMain: '#FEAC23',
   colorBg: '#212025',
+  colorNavBg: '#2F3136',
   colorWhite: '#FFFFFF',
-  colorC4: '#c4c4c4',
+  colorBorder: '#c4c4c4',
   colorDB: '#dbdbdb',
   color76: '#767676',
+  colorWarning: '#EB5757',
 };
 
 // const lightColors = {
